README.md: Clarify the name as relative to the authoritative_source
The name should be relative to the authoritative_source so that combination of the authoritative_source, name, version and release labels should resolve to the actual image (depending on the implementation details of the protocls used).

For Docker images created by Red Hat as well as for certified images, we standardized 'name' to be the relative path for the image (e.g. 'rhel7/rhel-tools'). An actual Red Hat image can be retrieved by:

    # docker pull authoritative_source+'/'+name+':'+version+'-'+release